// libs/database/src/enums/account-type.enum.ts
export enum AccountType {
  // Asset accounts (debit normal balance)
  ASSET_USER_WALLET = 'ASSET_USER_WALLET',
  ASSET_USER_HELD = 'ASSET_USER_HELD',
  ASSET_MERCHANT_WALLET = 'ASSET_MERCHANT_WALLET',
  ASSET_SETTLEMENT_PENDING = 'ASSET_SETTLEMENT_PENDING',
  ASSET_CASH = 'ASSET_CASH',
  ASSET_RECEIVABLES = 'ASSET_RECEIVABLES',

  // Liability accounts (credit normal balance)
  LIABILITY_USER_DEPOSITS = 'LIABILITY_USER_DEPOSITS',
  LIABILITY_MERCHANT_PAYABLE = 'LIABILITY_MERCHANT_PAYABLE',
  LIABILITY_REFUND_RESERVE = 'LIABILITY_REFUND_RESERVE',

  // Revenue accounts (credit normal balance)
  REVENUE_TRANSACTION_FEE = 'REVENUE_TRANSACTION_FEE',
  REVENUE_MONTHLY_FEE = 'REVENUE_MONTHLY_FEE',
  REVENUE_FX_SPREAD = 'REVENUE_FX_SPREAD',
  REVENUE_INTERCHANGE = 'REVENUE_INTERCHANGE',

  // Expense accounts (debit normal balance)
  EXPENSE_REFUND = 'EXPENSE_REFUND',
  EXPENSE_CHARGEBACK = 'EXPENSE_CHARGEBACK',
  EXPENSE_FRAUD_LOSS = 'EXPENSE_FRAUD_LOSS',
}

export enum LedgerEntryType {
  DEBIT = 'DEBIT',
  CREDIT = 'CREDIT',
}

export enum JournalType {
  AUTHORIZATION = 'AUTHORIZATION',
  AUTHORIZATION_RELEASE = 'AUTHORIZATION_RELEASE',
  CAPTURE = 'CAPTURE',
  SETTLEMENT = 'SETTLEMENT',
  REFUND = 'REFUND',
  P2P_TRANSFER = 'P2P_TRANSFER',
  TOPUP = 'TOPUP',
  WITHDRAWAL = 'WITHDRAWAL',
  FEE = 'FEE',
  ADJUSTMENT = 'ADJUSTMENT',
}